refactor(SPEC-CLI-TUI-MODERNIZE-001): M2 remove the dead legacy merge-confirmation TUI
internal/merge/confirm.go carried a second, parallel Bubble Tea program with
its own styling system (41 direct lipgloss usages, 4 raw hex colour values)
instead of consuming the internal/tui token source. Its entry point
ConfirmMerge had zero production callers — the live path routes update.go ->
confirmViaPreview -> update.PreviewClassification.

Reachability re-verified at execution time rather than carried from plan
authoring: `grep -rn ConfirmMerge` finds one call site and it is a test; the
cross-package census returns FileAnalysis 6 / MergeAnalysis 3 / ConfirmMerge 4
(all four comment-only); no reflection dispatch and no build-tag-gated caller
exists in internal/merge. The result matched the planned inventory exactly.

Removed: ConfirmMerge, confirmModel and its Init/Update/View/syncSelectedFiles/
filterSelectedFiles, fileListItem and its list adapter methods, the
AnalysisFormatter stack (formatterStyles, both constructors, initFormatterStyles,
every Format*/render*/format* method), truncateRowField, groupFilesByRisk,
groupByPathPrefix, pathGroup, topTwoSegments, riskIcon, validateAnalysis,
sanitizePath, and the maxMergeFiles/maxPathLength constants.

Preserved: MergeAnalysis and FileAnalysis, relocated verbatim to types.go
(the package's type home) with their field sets and exported names unchanged.
confirm.go is deleted outright rather than left as a 20-line file under a
filename that no longer describes its contents. Four production consumer sites
continue to resolve, including the field-set-sensitive named-field composite
literal in internal/cli/update/plan.

Test reconciliation went further than "delete both files": two tests in the
deleted files covered SURVIVING code and were retargeted rather than
discarded — TestMergeAnalysis_Creation (the preserved struct field sets) moved
to types_test.go alongside a new named-field-literal guard, and
TestToMapInterface moved to strategies_test.go, where the function it
exercises actually lives. Deleting it with the file would have silently
dropped coverage of live strategy code.

Security note (a finding, not a scope item): validateAnalysis (file-count and
path-length DoS limits) and sanitizePath (path-traversal sanitization) were
reachable only from the dead ConfirmMerge, so they were already not running in
production and removing them regresses no live control. They do surface that
the live path applies neither. That gap is pre-existing and out of scope here;
it is recorded so the removal does not erase the only trace of it.

The Windows non-TTY guard is unaffected: the live guard in update.go is
strictly broader (all platforms, not just Windows) and predates this change.

Measured: internal/merge coverage 86.3% -> 90.8%; terminal-UI library imports
in internal/merge 3 -> 0; raw hex colour literals on executable production
lines outside internal/tui 2 -> 0.

🗿 MoAI
